+++ apps/presentations/ui/presenter-notes-controller.js
@@ -99,7 +99,7 @@
   }
 
   global.InkDeskPresentationsNotes = Object.freeze({
-    version: '0.20.2.6',
+    version: '0.20.2.7',
     create(options) {
       return new PresentationNotesController(options);
     },